src/core/performance/cpu/vectorization.py
"""CPU Vectorization Framework for Adaptive Attention Tiling.

This module provides vectorized operations optimized for CPU execution,
focusing on attention computation, pattern dynamics, and geometric flows.
"""

import numpy as np
import torch
import torch.nn.functional as F
from typing import List, Tuple, Optional, Union, Callable
from dataclasses import dataclass
from functools import wraps

@dataclass
class VectorizationMetrics:
    """Metrics for vectorized operations."""
    execution_time: float
    memory_usage: float
    vectorization_efficiency: float
    operation_type: str

class VectorizationOptimizer:
    """Optimizes operations using CPU vectorization techniques."""
    
    def __init__(self, 
                 enable_profiling: bool = True,
                 use_mixed_precision: bool = True,
                 chunk_size: int = 1024):
        self.enable_profiling = enable_profiling
        self.use_mixed_precision = use_mixed_precision
        self.chunk_size = chunk_size
        self.metrics: List[VectorizationMetrics] = []

    def profile_vectorization(self, func: Callable) -> Callable:
        """Decorator to profile vectorized operations."""
        @wraps(func)
        def wrapper(*args, **kwargs):
            if not self.enable_profiling:
                return func(*args, **kwargs)
            
            start_time = torch.cuda.Event(enable_timing=True)
            end_time = torch.cuda.Event(enable_timing=True)
            
            start_time.record()
            result = func(*args, **kwargs)
            end_time.record()
            
            torch.cuda.synchronize()
            execution_time = start_time.elapsed_time(end_time)
            
            # Calculate memory usage
            memory_usage = torch.cuda.max_memory_allocated() / 1024**2  # MB
            
            # Estimate vectorization efficiency
            vectorization_efficiency = self._estimate_vectorization_efficiency(result)
            
            self.metrics.append(VectorizationMetrics(
                execution_time=execution_time,
                memory_usage=memory_usage,
                vectorization_efficiency=vectorization_efficiency,
                operation_type=func.__name__
            ))
            
            return result
        return wrapper

    def vectorize_attention(self, 
                          query: torch.Tensor, 
                          key: torch.Tensor, 
                          value: torch.Tensor,
                          mask: Optional[torch.Tensor] = None) -> torch.Tensor:
        """Vectorized attention computation."""
        # Ensure contiguous memory layout
        query, key, value = map(lambda x: x.contiguous(), (query, key, value))
        
        # Split into chunks for better cache utilization
        chunk_size = min(self.chunk_size, query.size(0))
        
        def attention_chunk(q: torch.Tensor, k: torch.Tensor, v: torch.Tensor) -> torch.Tensor:
            scores = torch.matmul(q, k.transpose(-2, -1)) / np.sqrt(k.size(-1))
            if mask is not None:
                scores = scores.masked_fill(mask == 0, float('-inf'))
            attn = F.softmax(scores, dim=-1)
            return torch.matmul(attn, v)
        
        # Use vmap for vectorized computation
        batched_attention = torch.vmap(attention_chunk, 
                                     in_dims=(0, 0, 0),
                                     out_dims=0)
        
        # Process in chunks
        outputs = []
        for i in range(0, query.size(0), chunk_size):
            chunk_q = query[i:i + chunk_size]
            chunk_k = key[i:i + chunk_size]
            chunk_v = value[i:i + chunk_size]
            outputs.append(batched_attention(chunk_q, chunk_k, chunk_v))
        
        return torch.cat(outputs, dim=0)

    def vectorize_attention(self, 
                          query: torch.Tensor,
                          key: torch.Tensor,
                          value: torch.Tensor) -> torch.Tensor:
        """Compute attention using vectorized operations."""
        # Scale dot-product attention
        d_k = query.size(-1)
        scores = torch.matmul(query, key.transpose(-2, -1)) / np.sqrt(d_k)
        attention = F.softmax(scores, dim=-1)
        
        # Apply attention to values
        output = torch.matmul(attention, value)
        
        if self.enable_profiling:
            memory_usage = sum(x.nelement() * x.element_size() for x in [query, key, value, scores, attention, output])
            vectorization_efficiency = torch.cuda.get_device_properties(0).multi_processor_count
            
            self.metrics.append(VectorizationMetrics(
                execution_time=0.0,  # Will be set by decorator
                memory_usage=memory_usage,
                vectorization_efficiency=vectorization_efficiency,
                operation_type="attention"
            ))
        
        return output

    def vectorize_pattern_dynamics(self, 
                                 pattern: torch.Tensor,
                                 flow: torch.Tensor) -> torch.Tensor:
        """Vectorized pattern dynamics computation."""
        def compute_dynamics(p: torch.Tensor, f: torch.Tensor) -> torch.Tensor:
            # Apply flow to pattern
            evolved = p + f * self._compute_evolution_rate(p)
            # Ensure stability
            return torch.clamp(evolved, -1, 1)
        
        # Use vmap for vectorized computation
        batched_dynamics = torch.vmap(compute_dynamics,
                                    in_dims=(0, 0),
                                    out_dims=0)
        
        return batched_dynamics(pattern, flow)

    def vectorize_geometric_flow(self,
                               metric: torch.Tensor,
                               connection: torch.Tensor) -> torch.Tensor:
        """Vectorized geometric flow computation."""
        def compute_flow(m: torch.Tensor, c: torch.Tensor) -> torch.Tensor:
            # Compute Ricci flow
            ricci = torch.einsum('...ij,...jk->...ik', m, c)
            return -2 * ricci  # Standard Ricci flow equation
        
        # Use vmap for vectorized computation
        batched_flow = torch.vmap(compute_flow,
                                 in_dims=(0, 0),
                                 out_dims=0)
        
        return batched_flow(metric, connection)

    def _compute_evolution_rate(self, pattern: torch.Tensor) -> torch.Tensor:
        """Compute evolution rate based on pattern characteristics."""
        energy = torch.sum(pattern**2, dim=-1, keepdim=True)
        return torch.exp(-energy) * 0.1  # Damping factor

    def _estimate_vectorization_efficiency(self, tensor: torch.Tensor) -> float:
        """Estimate vectorization efficiency based on memory layout and operations."""
        # Check memory layout efficiency
        memory_efficiency = 1.0 if tensor.is_contiguous() else 0.8
        
        # Check operation efficiency (simplified)
        op_efficiency = 0.9  # Assumed efficiency for vectorized operations
        
        return memory_efficiency * op_efficiency

    def get_metrics(self) -> List[VectorizationMetrics]:
        """Get collected vectorization metrics."""
        return self.metrics

    def clear_metrics(self) -> None:
        """Clear collected metrics."""
        self.metrics.clear()

    def clear_metrics(self) -> None:
        """Clear collected metrics."""
        self.metrics.clear()
